requirements.txt
asgiref==3.8.1
build==1.1.1
CacheControl==0.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
coverage==7.4.4
crashtest==0.4.1
distlib==0.3.8
Django==5.1.4
djangorestframework==3.15.2
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
idna==3.6
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.1
keyring==24.3.1
mccabe==0.7.0
more-itertools==10.2.0
msgpack==1.0.8
packaging==23.2
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-cov==5.0.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.3
tomlkit==0.12.4
trove-classifiers==2024.3.3
tzdata==2024.2
urllib3==2.2.1
virtualenv==20.25.1
